+++ frontend/.gitignore
@@ -28,11 +28,15 @@ jspm_packages/
 *.tgz
 .yarn-integrity
 .cache
+out
 !/.projenrc.js
 !/test/
 !/tsconfig.json
-!/src/
+!/pages/
 /lib
 /dist/
-/build/
-!/react-app-env.d.ts
+!/.eslintrc.json
+!/tsconfig.eslint.json
+!/tailwind.config.json
+!/postcss.config.json
+/.next/
