+++ frontend/.npmignore
@@ -1,7 +1,7 @@
 # ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
 /.projen/
 /test/
-/src/
+/pages/
 !/lib/
 !/lib/**/*.js
 !/lib/**/*.d.ts
@@ -12,4 +12,8 @@ dist
 /.idea/
 /.projenrc.js
 tsconfig.tsbuildinfo
-/build/
+/.eslintrc.json
+/tsconfig.eslint.json
+/tailwind.config.json
+/postcss.config.json
+/.next/
