+++ frontend/next-env.d.ts
@@ -1,6 +1,5 @@
 /// <reference types="next" />
 /// <reference types="next/types/global" />
 /// <reference types="next/image-types/global" />
-
 // NOTE: This file should not be edited
 // see https://nextjs.org/docs/basic-features/typescript for more information.
