+++ frontend/package.json
@@ -10,22 +10,28 @@
     "upgrade-projen": "npx projen upgrade-projen",
     "default": "npx projen default",
     "watch": "npx projen watch",
+    "eslint": "npx projen eslint",
     "dev": "npx projen dev",
-    "eject": "npx projen eject",
+    "export": "npx projen export",
+    "server": "npx projen server",
+    "telemetry": "npx projen telemetry",
     "projen": "npx projen"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.1.1",
-    "@testing-library/user-event": "^13.2.1",
-    "@types/jest": "^27.0.2",
-    "@types/node": "^14.17.0",
+    "@types/node": "^14.15.0",
     "@types/react": "^17.0.26",
     "@types/react-dom": "^17.0.9",
     "@types/react-scroll": "^1.8.3",
+    "@typescript-eslint/eslint-plugin": "^4.32.0",
+    "@typescript-eslint/parser": "^4.32.0",
+    "eslint": "^7.32.0",
+    "eslint-import-resolver-node": "^0.3.6",
+    "eslint-import-resolver-typescript": "^2.5.0",
+    "eslint-plugin-import": "^2.24.2",
+    "json-schema": "^0.3.0",
     "npm-check-updates": "^11",
     "projen": "^0.28.33",
-    "typescript": "^4.0.3"
+    "typescript": "4.3"
   },
   "dependencies": {
     "@headlessui/react": "^1.4.1",
@@ -34,35 +40,19 @@
     "@tailwindcss/forms": "^0.3.4",
     "@tailwindcss/line-clamp": "^0.2.1",
     "@tailwindcss/typography": "^0.4.1",
-    "autoprefixer": "^9",
-    "postcss": "^8",
+    "autoprefixer": "^10.3.6",
+    "next": "^11.1.2",
+    "postcss": "^8.3.8",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-scripts": "^4.0.0",
     "react-scroll": "^1.8.4",
-    "tailwindcss": "^2",
-    "web-vitals": "^2.1.0"
+    "tailwindcss": "^2.2.16"
   },
   "bundledDependencies": [],
+  "engines": {
+    "node": ">= 14.15.0"
+  },
   "license": "Apache-2.0",
   "version": "0.0.0",
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
   "//": "~~ Generated by projen. To modify, edit .projenrc.js and run \"npx projen\"."
 }
