+++ frontend/tailwind.config.json
@@ -1,0 +1,12 @@
+{
+  "purge": [],
+  "darkMode": false,
+  "theme": {
+    "extend": {}
+  },
+  "variants": {
+    "extend": {}
+  },
+  "plugins": [],
+  "//": "~~ Generated by projen. To modify, edit .projenrc.js and run \"npx projen\"."
+}
