+++ frontend/tsconfig.eslint.json
@@ -1,7 +1,5 @@
 {
   "compilerOptions": {
-    "rootDir": "pages",
-    "outDir": "lib",
     "alwaysStrict": true,
     "declaration": true,
     "experimentalDecorators": true,
@@ -36,10 +34,14 @@
     "noEmit": true
   },
   "include": [
+    ".projenrc.js",
     "pages/**/*.ts",
+    "test/**/*.ts",
     "**/*.ts",
     "**/*.tsx"
   ],
-  "exclude": [],
+  "exclude": [
+    "node_modules"
+  ],
   "//": "~~ Generated by projen. To modify, edit .projenrc.js and run \"npx projen\"."
 }
